requirements-dev.txt
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.2.1
defusedxml==0.7.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.22
django-allauth==0.44.0
django-bootstrap-form==3.4
django-debug-toolbar==3.1.1
idna==2.10
oauthlib==3.1.0
Pillow==7.2.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
static3==0.7.0
urllib3==1.26.4
